Change user borrowed books and delete exception

+++ server/controllers/users.js
@@ -1,4 +1,5 @@
 import User from '../models/user.js'
+import Book from '../models/book.js'
 
 export const getUsers = async (req, res) => {
   try {
@@ -11,8 +12,8 @@ export const getUsers = async (req, res) => {
 
 export const createUser = async (req, res) => {
   try {
-    const { name, email } = req.body
-    const newUser = await User.create({ name, email })
+    const { name, email, hasBorrowed } = req.body
+    const newUser = await User.create({ name, email, hasBorrowed })
     res.status(201).json({ message: 'User added', data: newUser })
   } catch (err) {
     res.status(500).json({ message: 'Error creating user' })
@@ -33,6 +34,8 @@ export const updateUser = async (req, res) => {
 export const deleteUser = async (req, res) => {
   try {
     const { id } = req.params
+    const hasBorrowed = await Book.findOne({ user: id })
+    if (hasBorrowed) throw new Error('User has borrowed a book')
     const data = await User.findByIdAndDelete(id)
     if (!data) throw new Error('User not found')
     res.status(200).json({ message: 'User deleted', data })
